user tests with authentication

+++ backend/tests/endpoints/user_test.py
@@ -36,20 +36,26 @@ def user_db_session():
     for table in reversed(db.metadata.sorted_tables):
         session.execute(table.delete())
     session.commit()
+
 class TestUserEndpoint:
     """Class to test user management endpoints."""
 
     def test_delete_user(self, client,user_db_session):
         """Test deleting a user."""
         # Delete the user
-        response = client.delete("/users/del")
+        response = client.delete("/users/del", headers={"Authorization":"Bearer del_user"})
         assert response.status_code == 200
         assert response.json["message"] ==  "User deleted successfully!"
 
     def test_delete_not_present(self, client,user_db_session):
         """Test deleting a user that does not exist."""
-        response = client.delete("/users/non")
-        assert response.status_code == 404
+        response = client.delete("/users/u_get", headers={"Authorization":"Bearer teacher1"})
+        assert response.status_code == 403 # User does not exist, so you are not the user
+
+    def test_delete_not_yourself(self, client, user_db_session):
+        """Test deleting a user that is not yourself."""
+        response = client.delete("/users/non", headers={"Authorization":"Bearer teacher1"})
+        assert response.status_code == 403 # User does not exist, so you are not the user
 
     def test_wrong_form_post(self, client,user_db_session):
         """Test posting with a wrong form."""
@@ -58,7 +64,7 @@ class TestUserEndpoint:
             'is_student': True,  # wrong field name
             'is_admin': False
         })
-        assert response.status_code == 400
+        assert response.status_code == 403 # Post not allowed
 
     def test_wrong_datatype_post(self, client,user_db_session):
         """Test posting with a wrong data type."""
@@ -67,18 +73,18 @@ class TestUserEndpoint:
             'is_teacher': True,
             'is_admin': False
         })
-        assert response.status_code == 415
+        assert response.status_code == 403 # Post not allowed
 
     def test_get_all_users(self, client,user_db_session):
         """Test getting all users."""
-        response = client.get("/users")
+        response = client.get("/users", headers={"Authorization":"Bearer teacher1"})
         assert response.status_code == 200
         # Check that the response is a list (even if it's empty)
         assert isinstance(response.json["data"], list)
 
     def test_get_one_user(self, client,user_db_session):
         """Test getting a single user."""
-        response = client.get("users/u_get")
+        response = client.get("users/u_get", headers={"Authorization":"Bearer teacher1"})
         assert response.status_code == 200
         assert response.json["data"] == {
             'uid': 'u_get',
@@ -92,8 +98,7 @@ class TestUserEndpoint:
             'is_teacher': False,
             'is_admin': True
         })
-        assert response.status_code == 200
-        assert response.json["message"] == "User updated successfully!"
+        assert response.status_code == 403 # Patching is not allowed
 
     def test_patch_non_existent(self, client,user_db_session):
         """Test updating a non-existent user."""
@@ -101,7 +106,7 @@ class TestUserEndpoint:
             'is_teacher': False,
             'is_admin': True
         })
-        assert response.status_code == 404
+        assert response.status_code == 403 # Patching is not allowed
 
     def test_patch_non_json(self, client,user_db_session):
         """Test sending a non-JSON patch request."""
@@ -110,12 +115,12 @@ class TestUserEndpoint:
             'is_teacher': True,
             'is_admin': False
         })
-        assert response.status_code == 415
+        assert response.status_code == 403 # Patching is not allowed
 
     def test_get_users_with_query(self, client, user_db_session):
         """Test getting users with a query."""
         # Send a GET request with query parameters, this is a nonsense entry but good for testing
-        response = client.get("/users?is_admin=true&is_teacher=false")
+        response = client.get("/users?is_admin=true&is_teacher=false", headers={"Authorization":"Bearer teacher1"})
         assert response.status_code == 200
 
         # Check that the response contains only the user that matches the query
@@ -124,3 +129,13 @@ class TestUserEndpoint:
         assert users[0]["uid"] == "query_user"
         assert users[0]["is_admin"] is True
         assert users[0]["is_teacher"] is False
+
+    def test_get_user_not_authenticated(self, client, user_db_session):
+        """Test getting a single user without authentication."""
+        response = client.get("users/u_get")
+        assert response.status_code == 401
+
+    def test_get_user_wrong_authenticated(self, client, user_db_session):
+        """Test getting a single user with incorrect authentication."""
+        response = client.get("users/u_get", headers={"Authorization":"Bearer wrong_token"})
+        assert response.status_code == 401
